+++ otj-server-reactive/src/main/java/com/opentable/server/reactive/ReactiveServer.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.opentable.server;
+package com.opentable.server.reactive;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -20,26 +20,25 @@ import java.lang.annotation.Target;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-
-import com.opentable.components.filterorder.FilterOrderResolverConfiguration;
+import org.springframework.web.reactive.config.DelegatingWebFluxConfiguration;
 
 /**
- * REST Reactive HTTP Server.
+ * Configures Spring WebFlux reactive server.
+ *
+ * Uses {@link DelegatingWebFluxConfiguration} to customize the configuration by delegating
+ * to beans of type {@link org.springframework.web.reactive.config.WebFluxConfigurer}, allowing them to
+ * customize the configuration provided by {@code WebFluxConfigurationSupport}.
+ *
+ * See {@link ReactiveServerUnconfigured} for more advanced control of configuration.
+ *
+ * This annotation is generally equivalent to {@link org.springframework.web.reactive.config.EnableWebFlux}.
  */
 @Configuration
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Import({
-    EmbeddedJettyConfiguration.class,
-    EmbeddedReactiveJetty.class,
-    JAXRSHttpServerCommonConfiguration.class,
-    // Filter for transfer core info to MDC
-    BackendInfoFilterConfiguration.class,
-    // Support static resources
-    StaticResourceConfiguration.class,
-    // Filter order
-    FilterOrderResolverConfiguration.class,
+        DelegatingWebFluxConfiguration.class,
 })
-@NonWebSetup
-public @interface RestReactiveHttpServer {
+@ReactiveServerUnconfigured
+public @interface ReactiveServer {
 }