+++ otj-server-reactive/src/test/java/com/opentable/server/reactive/AbstractTest.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.opentable.server.reactive;
+
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
+
+import javax.inject.Inject;
+import javax.inject.Provider;
+
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import com.opentable.server.HttpServerInfo;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = TestReactiveServerConfiguration.class, webEnvironment = RANDOM_PORT)
+@ActiveProfiles(profiles = "test")
+@TestPropertySource(properties = {
+        "info.component=test",
+})
+public abstract class AbstractTest {
+
+    HttpServerInfo httpServerInfo;
+
+    @Value("${local.server.port}")
+    int port;
+
+    @Inject
+    public void init(Provider<HttpServerInfo> info) {
+        this.httpServerInfo = info.get();
+    }
+}