+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 otj-server
 ==========
 
+3.0.7
+-----
+* Split Reactive Server in to its own top-level server component
+* Replace @RestReactiveHttpServer with @ReactiveServer annotation
+
 3.0.6
 -----
 * Fix EmbeddedReactiveJetty to work after module split (fixes regression post pom 180).